refactor: empty state 이미지 파일명 kebab-case로 통일
